Add service worker for offline capability
Works with the cache-first strategy, via sw-toolbox and broccoli-
service-worker.

Had to change paths to hash because the app was breaking on reload
whilst offline and in non-root route. This seems ok for now.
This means that path tests might be out of date; tests will have
to be reworked in general though (pouch fixtures, handle workers).

Future:
Should (?) add AppCache as fallback in the absence of sw.
Hospitalrun has a good implementation of this.

Add pouchDB syncing inside of service worker as well, to prevent
blocking and stutter.