Fetch current user from token API, add to session
Our API backend does not return a user object upon authenticating
rather a JWT token, per OAuth. In order to get the user object
from it we have to hit the /api/user/current endpoint, which reads the
user from the token and returns them formatted as JSONAPI.
We use ember-network fetch, because it is Fastboot-friendly.