Add basic sticker implementation
Add relationships in the user and sticker model, and functionality
to load the current user's stickers in a route model hook.

Also added functionality to send a sticker, but it is incomplete
in the absence of the receiver object. It works for a loaded user
(current), whose id we know though, so it is encouraging.

In the future, the sticker view should be a separate route,
as well as a card component implementation. The action should
also become a task.