Only load new stickers when on list edge
Again, the CORS bug prevents doing .currentUser.get('sticker') and
background reloading, so we have to be a bit more resourceful.
In this case, user stickers are reloaded when the list index
is the start or end, which minimises requests. There is a visual
annoyance with displaying "processing" for a button when the
loading time is so small, but overall seems to work well!