Add deferred loading pattern for stickers
Now instead of returning a big promise, we return all the data
already loaded (using .peek), plus an ember-concurrency Task that
is responsible for reloading the data. The template shows the
existing data first, and the new data once the taks finishes.
It also displays a loading spinner for the data it is still
fetching. This could supplant the model loading hook tbh.
Many thanks to Balint Erdi for demonstrating this pattern.